Modify RemoteDuinoServer to manage its buffer...
Modify RemoteDuinoServer to manage its own buffer.  Most importantly, modify
the RemoteDuinoServer parsing code to use a buffer of any size - this will
dramatically help with dynamic memory usage, since the web request doesn't need
to be loaded entirely into memory.  In fact, requests should be able to be
processed (albeit slowly) using a buffer of size as low as 1 character.

NOTE: currently the RemoteDuinoServer is only listening on the serial port.